refactor: consolidate baseline utils into baseline.ts
