+++ serving/mcp-server/data/baseline.ts
@@ -1,5 +1,4 @@
 import { features } from 'web-features';
-import { getFeatureStatus, mapBaseline, resolveFeatureId } from './baseline-utils.ts';
 
 export type BaselineStatus = 'Limited availability' | `Baseline since ${string}`;
 
@@ -19,18 +18,74 @@ export interface FeatureValidationResult {
 export interface DetailedBaselineStatus {
   baseline: 'low' | 'high' | false;
   baseline_low_date?: string;
+  baseline_high_date?: string;
+}
+
+/**
+ * Resolves the feature ID to its canonical form, following moves and splits.
+ */
+export function resolveFeatureId(featureId: string): string[] {
+  const feature = features[featureId] as Feature | undefined;
+  if (!feature) return [];
+  if (feature.kind === "feature") return [featureId];
+  if (feature.kind === "moved") return resolveFeatureId(feature.redirect_target);
+  if (feature.kind === "split") return feature.redirect_targets.flatMap(resolveFeatureId);
+  return [];
+}
+
+/**
+ * Gets the detailed Baseline status for a specific feature, resolving redirects/splits.
+ */
+export function getFeatureStatus(featureId: string): DetailedBaselineStatus | undefined {
+  const resolvedIds = resolveFeatureId(featureId);
+  if (resolvedIds.length === 0) return;
+
+  let latestLowDate = "0000-00-00";
+  let latestHighDate = "0000-00-00";
+  let overallBaseline: 'low' | 'high' | false = 'high';
+
+  for (const id of resolvedIds) {
+    const feature = features[id] as Feature;
+    if (feature.kind !== 'feature' || !feature.status) {
+      overallBaseline = false;
+      continue;
+    }
+    
+    const status = feature.status;
+    if (status.baseline === false) overallBaseline = false;
+    else if (status.baseline === 'low' && overallBaseline === 'high') overallBaseline = 'low';
+
+    if (status.baseline_low_date && status.baseline_low_date > latestLowDate) {
+      latestLowDate = status.baseline_low_date;
+    }
+    if (status.baseline_high_date && status.baseline_high_date > latestHighDate) {
+      latestHighDate = status.baseline_high_date;
+    }
+  }
+
+  return {
+    baseline: overallBaseline,
+    baseline_low_date: latestLowDate === "0000-00-00" ? undefined : latestLowDate,
+    baseline_high_date: latestHighDate === "0000-00-00" ? undefined : latestHighDate
+  };
 }
 
 /**
  * Gets the detailed Baseline status for a specific feature.
+ * (Legacy wrapper for getFeatureStatus)
  */
 export function getDetailedBaselineStatus(featureId: string): DetailedBaselineStatus | undefined {
-  const status = getFeatureStatus(featureId);
-  if (!status) return;
-  return {
-    baseline: status.baseline,
-    baseline_low_date: status.baseline_low_date
-  };
+  return getFeatureStatus(featureId);
+}
+
+/**
+ * Maps baseline internal values to human-readable terms.
+ */
+export function mapBaseline(baseline: string | boolean | undefined): string {
+  if (baseline === 'low') return 'Newly available';
+  if (baseline === 'high') return 'Widely available';
+  if (baseline === false) return 'Limited availability';
+  return 'unknown';
 }
 
 /**
@@ -77,11 +132,6 @@ export function checkBaseline(target: string, featureId: string): boolean {
 
   return false;
 }
-
-/**
- * Resolves the feature ID to its canonical form.
- */
-export { resolveFeatureId } from './baseline-utils.ts';
 
 /**
  * Validates a feature ID.
@@ -142,7 +192,7 @@ export function getStatusMessage(featureId: string, bcdKey?: string): string | u
   const feature = features[featureId] as Feature | undefined;
   if (!feature) return;
 
-  const baselineStatus = getDetailedBaselineStatus(featureId);
+  const baselineStatus = getFeatureStatus(featureId);
   if (!baselineStatus) return;
 
   const subject = (feature as any).name || featureId;
